requirements.txt
Django==4.2.14
pdf2image==1.17.0
Pillow==9.4.0
pytesseract~=0.3.10
reportlab~=4.2.2
requests~=2.29.0
pathlib~=1.0.1
pdfplumber~=0.11.3